refactor or_* table entities db/or/ with table3/fnHelper
