+++ convex/openrouter/entities/providers.ts
@@ -1,121 +1,7 @@
-import { asyncMap } from 'convex-helpers'
-import { v } from 'convex/values'
-
-import { diff as jsonDiff, type IChange } from 'json-diff-ts'
-
-import { internalMutation, query, type MutationCtx } from '../../_generated/server'
-import { Table2 } from '../../table2'
-import { countResults } from '../output'
-
-export const OrProviders = Table2('or_providers', {
-  slug: v.string(),
-  name: v.string(),
-  headquarters: v.optional(v.string()),
-  datacenters: v.optional(v.array(v.string())),
-  icon: v.object({
-    url: v.string(),
-    invertRequired: v.optional(v.boolean()),
-  }),
-  status_page_url: v.optional(v.string()),
-  moderation_required: v.boolean(),
-
-  capabilities: v.object({
-    completions: v.boolean(),
-    chat_completions: v.boolean(),
-    multipart_messages: v.boolean(),
-    stream_cancellation: v.boolean(),
-    byok: v.boolean(),
-  }),
-
-  data_policy: v.object({
-    terms_of_service_url: v.optional(v.string()),
-    privacy_policy_url: v.optional(v.string()),
-    data_policy_url: v.optional(v.string()),
-    requires_user_ids: v.optional(v.boolean()),
-
-    paid_models: v.object({
-      training: v.boolean(),
-      retains_prompts: v.optional(v.boolean()),
-      retention_days: v.optional(v.number()),
-      can_publish: v.optional(v.boolean()),
-    }),
-
-    free_models: v.optional(
-      v.object({
-        training: v.boolean(),
-        retains_prompts: v.optional(v.boolean()),
-        retention_days: v.optional(v.number()),
-        can_publish: v.optional(v.boolean()),
-      }),
-    ),
-  }),
-
-  snapshot_at: v.number(),
-})
-
-export const OrProvidersChanges = Table2('or_providers_changes', {
-  slug: v.string(),
-  snapshot_at: v.number(),
-  changes: v.array(v.record(v.string(), v.any())),
-})
-
-const diff = (a: unknown, b: unknown) =>
-  jsonDiff(a, b, {
-    keysToSkip: ['_id', '_creationTime', 'snapshot_at'],
-    embeddedObjKeys: {
-      datacenters: '$value',
-    },
-  })
-
-const recordChanges = async (
-  ctx: MutationCtx,
-  { content, changes }: { content: { slug: string; snapshot_at: number }; changes: IChange[] },
-) => {
-  if (changes.length === 0) return
-  const { slug, snapshot_at } = content
-  await ctx.db.insert(OrProvidersChanges.name, { slug, snapshot_at, changes })
-}
-
-export const upsert = internalMutation({
-  args: {
-    items: v.array(OrProviders.content),
-  },
-  handler: async (ctx, { items }) => {
-    const results = await asyncMap(items, async (item) => {
-      const existing = await ctx.db
-        .query(OrProviders.name)
-        .withIndex('by_slug', (q) => q.eq('slug', item.slug))
-        .first()
-      const changes = diff(existing ?? {}, item)
-
-      // Record changes
-      await recordChanges(ctx, { content: item, changes })
-
-      // Insert
-      if (!existing) {
-        await ctx.db.insert(OrProviders.name, item)
-        return { action: 'insert' }
-      }
-
-      // Stable - no changes
-      if (changes.length === 0) {
-        await ctx.db.patch(existing._id, { snapshot_at: item.snapshot_at })
-        return { action: 'stable' }
-      }
-
-      // Update
-      await ctx.db.replace(existing._id, item)
-      return { action: 'update' }
-    })
-
-    return countResults(results, 'providers')
-  },
-})
+import * as ORProviders from '../../db/or/providers'
 
 // * queries
+export const list = ORProviders.list.query
 
-export const list = query({
-  handler: async (ctx) => {
-    return await ctx.db.query('or_providers').collect()
-  },
-})
+// * snapshots
+export const upsert = ORProviders.upsert.internalMutation
