+++ convex/openrouter/pipelines/endpoints.ts
@@ -2,10 +2,10 @@ import type { Infer } from 'convex/values'
 
 import { internal } from '../../_generated/api'
 import type { ActionCtx } from '../../_generated/server'
+import * as OREndpoints from '../../db/or/endpoints'
+import { type EndpointStat } from '../../db/or/endpointStats'
 import * as ORModels from '../../db/or/models'
 import { storeSnapshotData } from '../archive'
-import { OrEndpoints } from '../entities/endpoints'
-import { type EndpointStat } from '../entities/endpointStats'
 import { validateArray, validateRecord, type Issue } from '../validation'
 import { EndpointStrictSchema, EndpointTransformSchema } from '../validators/endpoints'
 import {
@@ -27,6 +27,7 @@ type NewEndpointUptime = {
     uptime?: number
   }[]
 }
+
 export async function endpointsPipeline(
   ctx: ActionCtx,
   {
@@ -45,7 +46,7 @@ export async function endpointsPipeline(
   },
 ) {
   const started_at = Date.now()
-  const endpoints: (typeof OrEndpoints.$content)[] = []
+  const endpoints: Infer<typeof OREndpoints.vTable.validator>[] = []
   const endpointStats: NewEndpointStat[] = []
   const endpointUptimes: NewEndpointUptime[] = []
   const issues: Issue[] = []
